pyproject.toml
[tool.poetry]
name = "osxmetadata"
version = "0.99.38"
description = "Read and write meta data, such as tags/keywords, Finder comments, etc. on MacOS files"
authors = ["Rhet Turnbull <rturnbull+git@gmail.com>"]
license = "MIT License"
readme = "README.md"
homepage = "https://github.com/RhetTbull/osxmetadata"
repository = "https://github.com/RhetTbull/osxmetadata"
keywords = [
    "cli",
    "mac",
    "macos",
    "metadata",
    "tags",
    "osx",
    "spotlight",
    "extended_attributes",
]

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
bitstring = "^3.1.9"
click = "^8.0"
py-applescript = "^1.0.3"
pyobjc-core = "^8.5.1"
pyobjc-framework-AppleScriptKit = "^8.5.1"
pyobjc-framework-AppleScriptObjC = "^8.5.1"
pyobjc-framework-AVFoundation = "^8.5.1"
pyobjc-framework-CoreMedia = "^8.5.1"
pyobjc-framework-Quartz = "^8.5.1"
twine = "^4.0.1"
wheel = "^0.37.1"
xattr = "^0.9.9"
pyobjc-framework-CoreServices = "^8.5.1"
pyobjc-framework-ScriptingBridge = "^8.5.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.1.3"
build = "^0.8.0"
doit = "^0.36.0"
pyinstaller = "^5.4.1"
bump2version = "^1.0.1"

[tool.poetry.scripts]
osxmetadata = "osxmetadata.__main__:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
